requirements.txt
jinja2==3.1.4
pyyaml==6.0.2
checkov==3.2.46
tflint==0.51.0
unittest-xml-reporting==3.2.0
